requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
Flask-Migrate==4.0.7
marshmallow==3.23.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.4.2
python-dotenv==1.0.1
gunicorn==23.0.0

# Testing
pytest==8.3.3
pytest-flask==1.3.0
factory-boy==3.3.1
